Unify Rust bridge to zero-copy numpy arrays; rename transforms.rs to outline.rs
- render.rs: replace bytes() bridge with PyReadonlyArray1<i64/f32>; update
  command type comparisons from i32 to i64 to match PyTorch int64 tensors
- transforms.py: update render_bitmap to use the new numpy bridge
- Rename src/transforms.rs → src/outline.rs for naming consistency with
  render.rs (both are specific to their concern rather than generic)
- _torchfont.pyi: update signatures to reflect numpy array parameters
