+++ torchfont/_torchfont.pyi
@@ -1,6 +1,9 @@
 from collections.abc import Sequence
 
-def render_bitmap(types_bytes: bytes, coords_bytes: bytes, size: int) -> bytes: ...
+import numpy as np
+
+def render_bitmap(types: np.ndarray, coords: np.ndarray, size: int) -> bytes: ...
+def quad_to_cubic_inplace(types: np.ndarray, coords: np.ndarray, seq_len: int) -> None: ...
 
 class GlyphItem:
     types: bytes
